demo/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
numpy
pillow
torch
torchvision
transformers
